ecommerce_voicebot/backend/requirements.txt
python-dotenv
livekit
cartesia
livekit-agents
flask
flask-cors
livekit-agents[openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
chromadb
sentence-transformers

